add optional login mock mode for tests localhost fe <-> preview be
