+++ webapp/src/main/resources/application-preview.yml
@@ -50,6 +50,7 @@ management:
 
 security:
   secure-cookie: true
+  mock-mode: true
 
 metal-release-butler:
   host: http://butler-app:8080
